+++ rust/operator-binary/src/discovery.rs
@@ -4,6 +4,7 @@ use snafu::{OptionExt, ResultExt, Snafu};
 use stackable_opa_crd::{OpaCluster, OpaRole};
 use stackable_operator::{
     builder::{ConfigMapBuilder, ObjectMetaBuilder},
+    commons::product_image_selection::ResolvedProductImage,
     k8s_openapi::api::core::v1::{ConfigMap, Service},
     kube::{runtime::reflector::ObjectRef, Resource, ResourceExt},
 };
@@ -19,8 +20,6 @@ pub enum Error {
     NoName,
     #[snafu(display("object has no namespace associated"))]
     NoNamespace,
-    #[snafu(display("object has no version associated"))]
-    NoVersion { source: stackable_opa_crd::Error },
     #[snafu(display("failed to build ConfigMap"))]
     BuildConfigMap {
         source: stackable_operator::error::Error,
@@ -31,10 +30,17 @@ pub enum Error {
 pub fn build_discovery_configmaps(
     owner: &impl Resource<DynamicType = ()>,
     opa: &OpaCluster,
+    resolved_product_image: &ResolvedProductImage,
     svc: &Service,
 ) -> Result<Vec<ConfigMap>, Error> {
     let name = owner.name_any();
-    Ok(vec![build_discovery_configmap(&name, owner, opa, svc)?])
+    Ok(vec![build_discovery_configmap(
+        &name,
+        owner,
+        opa,
+        resolved_product_image,
+        svc,
+    )?])
 }
 
 /// Build a discovery [`ConfigMap`] containing information about how to connect to a certain [`OpaCluster`]
@@ -42,6 +48,7 @@ fn build_discovery_configmap(
     name: &str,
     owner: &impl Resource<DynamicType = ()>,
     opa: &OpaCluster,
+    resolved_product_image: &ResolvedProductImage,
     svc: &Service,
 ) -> Result<ConfigMap, Error> {
     let url = format!(
@@ -64,7 +71,7 @@ fn build_discovery_configmap(
                 })?
                 .with_recommended_labels(build_recommended_labels(
                     opa,
-                    opa.image_version().context(NoVersionSnafu)?,
+                    &resolved_product_image.app_version_label,
                     &OpaRole::Server.to_string(),
                     "discovery",
                 ))